Feature/blog auth

+++ blog-app-apis/src/main/java/com/API/blog/controllers/UserController.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,9 +16,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.API.blog.entities.Role;
+import com.API.blog.entities.User;
+import com.API.blog.exceptions.ResourceNotFoundException;
 import com.API.blog.payloads.ApiResponse;
 import com.API.blog.payloads.UserDto;
+import com.API.blog.repositories.RoleRepo;
+import com.API.blog.security.JwtAuthenticationFilter;
 import com.API.blog.services.UserService;
+import com.API.blog.utilts.JWTRequest;
 
 import jakarta.validation.Valid;
 
@@ -27,7 +34,10 @@ public class UserController {
 
 	@Autowired
 	private UserService userService;
-	
+	@Autowired
+	private JwtAuthenticationFilter jwtRequest;
+	@Autowired
+	private RoleRepo rolerepo;
 	//Post -create user
 	@PostMapping("/")
 	public ResponseEntity<UserDto> createUser(@Valid @RequestBody UserDto userDto){
@@ -45,9 +55,15 @@ public class UserController {
 	}
 	
 	// Delete user
-	
+
 	@DeleteMapping("/{userId}")
 	public ResponseEntity<ApiResponse> deleteUser (@PathVariable Integer userId) {
+		UserDto gettingUser = this.userService.getUserByID(userId);
+		if (gettingUser.getEmail().equals(jwtRequest.getUsername())) {
+			return new ResponseEntity<ApiResponse>(new ApiResponse("Can not delete existing user" , false) , HttpStatus.OK);
+		}
+		
+		
 		this.userService.deleteUser(userId);
 		return new ResponseEntity<ApiResponse>(new ApiResponse("User Deleted Successfully " , true) , HttpStatus.OK);
 	}
